+++ src/core/wallet.ts
@@ -1,29 +1,20 @@
 import { base64, hex } from "@scure/base";
 import { sha256 } from "@noble/hashes/sha256";
 import * as btc from "@scure/btc-signer";
+import { TAP_LEAF_VERSION, tapLeafHash } from "@scure/btc-signer/payment";
+import { secp256k1 } from "@noble/curves/secp256k1";
+import { clearInterval, setInterval } from "timers";
 
-import type {
-    Wallet as IWallet,
-    WalletConfig,
-    WalletBalance,
-    SendBitcoinParams,
-    AddressInfo,
-    Coin,
-    VirtualCoin,
-    Identity,
-    SettleParams,
-    VtxoTaprootAddress,
-    SpendableVtxo,
-} from "../types/wallet";
-import { ESPLORA_URL, EsploraProvider } from "../providers/esplora";
-import { ArkProvider } from "../providers/ark";
 import { BIP21 } from "../utils/bip21";
 import { ArkAddress } from "../core/address";
 import { checkSequenceVerifyScript, VtxoTapscript } from "../core/tapscript";
 import { selectCoins, selectVirtualCoins } from "../utils/coinselect";
-import { getNetwork, Network, NetworkName } from "../types/networks";
-import { TAP_LEAF_VERSION, tapLeafHash } from "@scure/btc-signer/payment";
-import { secp256k1 } from "@noble/curves/secp256k1";
+import { getNetwork, Network, NetworkName } from "./networks";
+import {
+    ESPLORA_URL,
+    EsploraProvider,
+    OnchainProvider,
+} from "../providers/onchain";
 import {
     ArkInfo,
     FinalizationEvent,
@@ -31,13 +22,17 @@ import {
     SettlementEventType,
     SigningNoncesGeneratedEvent,
     SigningStartEvent,
-} from "../providers/base";
-import { clearInterval, setInterval } from "timers";
+    ArkProvider,
+    Output,
+    VtxoInput,
+    RestArkProvider,
+} from "../providers/ark";
 import { TreeSignerSession } from "./signingSession";
 import { buildForfeitTx } from "./forfeit";
 import { TxWeightEstimator } from "../utils/txSizeEstimator";
 import { validateConnectorsTree, validateVtxoTree } from "./tree/validation";
 import { TransactionOutput } from "@scure/btc-signer/psbt";
+import { Identity } from "./identity";
 
 const ZERO_32 = new Uint8Array([
     0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
@@ -45,10 +40,104 @@ const ZERO_32 = new Uint8Array([
     0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
 ]);
 
-export class Wallet implements IWallet {
+export interface WalletConfig {
+    network: NetworkName;
+    identity: Identity;
+    esploraUrl?: string;
+    arkServerUrl?: string;
+    arkServerPublicKey?: string;
+}
+
+export interface WalletBalance {
+    onchain: {
+        confirmed: number;
+        unconfirmed: number;
+        total: number;
+    };
+    offchain: {
+        swept: number;
+        settled: number;
+        pending: number;
+        total: number;
+    };
+    total: number;
+}
+
+export interface SendBitcoinParams {
+    address: string;
+    amount: number;
+    feeRate?: number;
+    memo?: string;
+}
+
+export interface Recipient {
+    address: string;
+    amount: number;
+}
+
+// SpendableVtxo embed the forfeit script to use as spending path for the boarding utxo or vtxo
+export type SpendableVtxo = VtxoInput & {
+    forfeitScript: string;
+};
+
+export interface SettleParams {
+    inputs: (string | SpendableVtxo)[];
+    outputs: Output[];
+}
+
+// VtxoTaprootAddress embed the tapscripts composing the address
+// it admits the internal key is the unspendable x-only public key
+export interface VtxoTaprootAddress {
+    address: string;
+    scripts: {
+        exit: string[];
+        forfeit: string[];
+    };
+}
+
+export interface AddressInfo {
+    onchain: string;
+    offchain?: VtxoTaprootAddress;
+    boarding?: VtxoTaprootAddress;
+    bip21: string;
+}
+
+export interface TapscriptInfo {
+    offchain?: string[];
+    boarding?: string[];
+}
+
+export interface Status {
+    confirmed: boolean;
+    block_height?: number;
+    block_hash?: string;
+    block_time?: number;
+}
+
+export interface VirtualStatus {
+    state: "pending" | "settled" | "swept" | "spent";
+    batchTxID?: string;
+    batchExpiry?: number;
+}
+
+export interface Outpoint {
+    txid: string;
+    vout: number;
+}
+
+export interface Coin extends Outpoint {
+    value: number;
+    status: Status;
+}
+
+export interface VirtualCoin extends Coin {
+    virtualStatus: VirtualStatus;
+}
+
+export class Wallet {
     private identity: Identity;
     private network: Network;
-    private onchainProvider: EsploraProvider;
+    private onchainProvider: OnchainProvider;
     private arkProvider?: ArkProvider;
     private unsubscribeEvents?: () => void;
     private onchainAddress: string;
@@ -77,10 +166,7 @@ export class Wallet implements IWallet {
 
         // Setup Ark provider and derive offchain address if configured
         if (config.arkServerUrl && config.arkServerPublicKey) {
-            this.arkProvider = new ArkProvider(
-                config.arkServerUrl,
-                config.arkServerPublicKey
-            );
+            this.arkProvider = new RestArkProvider(config.arkServerUrl);
 
             // Generate tapscripts for Offchain and Boarding address
             const serverPubKey = hex.decode(config.arkServerPublicKey);
